requirements.txt
Flask==2.3.2
Werkzeug<3.0.0
gunicorn==21.2.0
google-cloud-firestore==2.21.0
google-auth==2.23.4
google-auth-oauthlib==1.2.0
google-cloud-storage==2.14.0
earthengine-api==0.1.395
fastapi
uvicorn
earthengine-api
google-auth
google-auth-httplib2
google-cloud-storage
httplib2
requests
flask-cors
mysql-connector-python
